+++ NEWS.md
@@ -1,3 +1,8 @@
-# ipmisc 0.1.0
+# ipmisc 1.1.0.9000
 
-* Initial release of the package.
+  - Re-export `tibble::as_tibble` and color functions from `crayon`.
+  - New function: `easystats_to_tidy_names`, `sort_xy`.
+
+# ipmisc 1.0.0
+
+  - Initial release of the package.